# contracts/openapi.yaml
openapi: 3.1.0
info:
  title: StratusCloud Public API
  version: 0.3.0
  description: |
    StratusCloud 1차 MVP 공개 REST API 계약서.
    Week 3에서는 IAM-2 범위(OIDC JWT/API Key/Policy)를 포함한다.
servers:
  - url: https://api.stratuscloud.local
    description: local development gateway
security:
  - bearerAuth: []
  - apiKeyAuth: []
...
paths:
...
  /v1/projects/{projectId}:
    get:
      tags: [iam]
      summary: 프로젝트 단건 정보를 조회한다.
      operationId: getProject
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: 조회 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Project'
  /v1/iam/users:
    post:
      tags: [iam]
      summary: 사용자를 생성한다.
      operationId: createUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateUserRequest'
      responses:
        '201':
          description: 생성 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
  /v1/projects/{projectId}/members:
    post:
      tags: [iam]
      summary: 프로젝트 멤버를 추가한다.
      operationId: addProjectMember
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AddMembershipRequest'
      responses:
        '201':
          description: 추가 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Membership'
  /v1/projects/{projectId}/members/{userId}/role:
    patch:
      tags: [iam]
      summary: 프로젝트 멤버 역할을 변경한다.
      operationId: updateProjectMemberRole
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: userId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateMembershipRoleRequest'
      responses:
        '200':
          description: 변경 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Membership'
  /v1/iam/policies:
    post:
      tags: [iam]
      summary: 정책 문서를 생성한다.
      operationId: createPolicy
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreatePolicyRequest'
      responses:
        '201':
          description: 생성 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Policy'
    get:
      tags: [iam]
      summary: 테넌트의 정책 목록을 조회한다.
      operationId: listPolicies
      parameters:
        - name: tenantId
          in: query
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: 조회 성공
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Policy'
...
  /v1/iam/api-keys:
    post:
      tags: [iam]
      summary: API Key를 발급한다.
      operationId: issueApiKey
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateApiKeyRequest'
      responses:
        '201':
          description: 발급 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiKey'
    get:
      tags: [iam]
      summary: 테넌트의 API Key 목록을 조회한다.
      operationId: listApiKeys
      parameters:
        - name: tenantId
          in: query
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: 조회 성공
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ApiKey'
  /v1/iam/api-keys/{keyId}:
    delete:
      tags: [iam]
      summary: API Key를 회수한다.
      operationId: revokeApiKey
      parameters:
        - name: keyId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: 회수 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiKey'
...
  /v1/audit/logs:
    get:
      tags: [audit]
      summary: 감사로그를 필터 조건으로 조회한다.
      operationId: listAuditLogs
      parameters:
        - name: tenantId
          in: query
          required: true
          schema:
            type: string
            format: uuid
        - name: projectId
          in: query
          required: false
          schema:
            type: string
            format: uuid
            nullable: true
        - name: actorId
          in: query
          required: false
          schema:
            type: string
            format: uuid
        - name: resourceType
          in: query
          required: false
          schema:
            type: string
        - name: action
          in: query
          required: false
          schema:
            type: string
        - name: result
          in: query
          required: false
          schema:
            $ref: '#/components/schemas/AuditResult'
      responses:
        '200':
          description: 조회 성공
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/AuditLog'
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: OIDC issuer/JWKS 검증을 통과한 JWT 토큰
    apiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: 머신 간 호출용 API Key
  schemas:
...
    RoleType:
      type: string
      enum: [OWNER, ADMIN, DEVELOPER, VIEWER]
    ApiKeyStatus:
      type: string
      enum: [ACTIVE, REVOKED]
...
    AuditResult:
      type: string
      enum: [SUCCESS, DENIED]
...
    Project:
      type: object
      required: [id, tenantId, name, createdAt]
      properties:
        id:
          type: string
          format: uuid
        tenantId:
          type: string
          format: uuid
        name:
          type: string
        createdAt:
          type: string
          format: date-time
    CreateUserRequest:
      type: object
      required: [email, displayName]
      properties:
        email:
          type: string
          format: email
        displayName:
          type: string
          minLength: 2
          maxLength: 100
    User:
      type: object
      required: [id, email, displayName, createdAt]
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
          format: email
        displayName:
          type: string
        createdAt:
          type: string
          format: date-time
    AddMembershipRequest:
      type: object
      required: [userId, role]
      properties:
        userId:
          type: string
          format: uuid
        role:
          $ref: '#/components/schemas/RoleType'
    UpdateMembershipRoleRequest:
      type: object
      required: [role]
      properties:
        role:
          $ref: '#/components/schemas/RoleType'
    Membership:
      type: object
      required: [id, tenantId, projectId, userId, role, createdAt]
      properties:
        id:
          type: string
          format: uuid
        tenantId:
          type: string
          format: uuid
        projectId:
          type: string
          format: uuid
        userId:
          type: string
          format: uuid
        role:
          $ref: '#/components/schemas/RoleType'
        createdAt:
          type: string
          format: date-time
    PolicyDocument:
      type: object
      required: [version, statements]
      properties:
        version:
          type: string
        statements:
          type: array
          items:
            type: object
            required: [effect, actions, resources]
            properties:
              effect:
                type: string
                enum: [ALLOW, DENY]
              actions:
                type: array
                items:
                  type: string
              resources:
                type: array
                items:
                  type: string
    CreatePolicyRequest:
      type: object
      required: [tenantId, name, document]
      properties:
        tenantId:
          type: string
          format: uuid
        name:
          type: string
          minLength: 2
          maxLength: 100
        description:
          type: string
          maxLength: 300
        document:
          $ref: '#/components/schemas/PolicyDocument'
    Policy:
      type: object
      required: [id, tenantId, name, document, createdAt]
      properties:
        id:
          type: string
          format: uuid
        tenantId:
          type: string
          format: uuid
        name:
          type: string
        description:
          type: string
          nullable: true
        document:
          $ref: '#/components/schemas/PolicyDocument'
        createdAt:
          type: string
          format: date-time
...
    CreateApiKeyRequest:
      type: object
      required: [tenantId, name, role]
      properties:
        tenantId:
          type: string
          format: uuid
        projectId:
          type: string
          format: uuid
          nullable: true
        name:
          type: string
          minLength: 2
          maxLength: 100
        role:
          $ref: '#/components/schemas/RoleType'
        expiresAt:
          type: string
          format: date-time
          nullable: true
    ApiKey:
      type: object
      required: [id, tenantId, name, role, keyPrefix, status, expiresAt, createdAt]
      properties:
        id:
          type: string
          format: uuid
        tenantId:
          type: string
          format: uuid
        projectId:
          type: string
          format: uuid
          nullable: true
        name:
          type: string
        role:
          $ref: '#/components/schemas/RoleType'
        keyPrefix:
          type: string
        status:
          $ref: '#/components/schemas/ApiKeyStatus'
        expiresAt:
          type: string
          format: date-time
        revokedAt:
          type: string
          format: date-time
          nullable: true
        createdAt:
          type: string
          format: date-time
        rawKey:
          type: string
          nullable: true
          description: 발급 시 1회만 반환되는 원문 키
...
    AuditLog:
      type: object
      required: [id, traceId, actorId, tenantId, action, resourceType, result, metadata, occurredAt]
      properties:
        id:
          type: string
          format: uuid
        traceId:
          type: string
        actorId:
          type: string
          format: uuid
        tenantId:
          type: string
          format: uuid
        projectId:
          type: string
          format: uuid
          nullable: true
        action:
          type: string
        resourceType:
          type: string
        resourceId:
          type: string
          nullable: true
        result:
          $ref: '#/components/schemas/AuditResult'
        metadata:
          type: object
          additionalProperties: true
        occurredAt:
          type: string
          format: date-time
...